Move the hex unit test to primitives
`format` is already in scope. No other changes to the test.

Note this does introduce some duplication in the unit tests, leaving
refactor for later.
